requirements-test.txt
# Testing framework
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1
pytest-mock==3.12.0
# Code quality tools
flake8==6.1.0
pylint==3.0.3
pydocstyle==6.3.0
lizard==1.17.10
bandit==1.7.5
isort==5.12.0
black==23.12.0
mypy==1.7.1
pre-commit==3.6.0
# Application dependencies
fastapi==0.104.1
uvicorn==0.24.0
requests==2.31.0
httpx==0.25.2
python-dotenv==1.0.0
# Utilities
responses==0.24.1
boto3
